pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "fastapi-cli"                      # nom du package (PyPI)
version = "0.1.0"
description = "A CLI tool for FastAPI projects"
authors = [
    { name="yann-arthur Tcheumani", email="tcheumaniyannarthur@proton.me" }
]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "typer>=0.9.0",
    "uvicorn>=0.20.0",
    "fastapi>=0.110.0",
    "SQLAlchemy>=2.0.30",
    "pydantic-settings>=2.2.1",
    "PyYAML>=6.0.2",
    "InquirerPy>=0.3.4",
    "dacite>=1.9.2"
]
license = { text = "MIT" }

[project.urls]
"Homepage" = "https://github.com/yanntcheumani/fastapi-cli"
"Bug Tracker" = "https://github.com/yanntcheumani/fastapi-cli/issues"

[project.scripts]
fastapi-cli = "fastapi_cli.cli:main"


[tool.setuptools.package-data]
fastapi_cli = ["default_archi/**/*"]

# [tool.setuptools.packages.find]
# where = ["fastapi_cli"]